+++ backend/src/docs/asciidoc/auth.adoc
@@ -2,6 +2,7 @@
 == Auth API
 
 === 로그인
+
 ==== 발생 가능 에러
 
 - 40002, 50001
