+++ backend/src/docs/asciidoc/inventoryProducts.adoc
@@ -2,6 +2,7 @@
 == InventoryProduct API
 
 === 나의 로그인 정보로 인벤토리를 조회
+
 ==== 발생 가능 에러
 
 - 40100, 40101, 40102
