+++ backend/src/docs/asciidoc/members.adoc
@@ -2,6 +2,7 @@
 == Member API
 
 === 로그인된 상태로 나의 회원 정보 조회
+
 ==== 발생 가능 에러
 
 - 40100, 40101, 40102
