+++ backend/src/main/java/com/woowacourse/f12/exception/ErrorCode.java
@@ -17,7 +17,7 @@ public enum ErrorCode {
     INVALID_REVIEW_CONTENT_LENGTH("40032"),
     ALREADY_WRITTEN_REVIEW("40033"),
 
-    PROFILE_CONTAINS_SOFTWARE("40040"),
+    INVALID_CATEGORY_PROFILE_PRODUCT("40040"),
     CATEGORY_DUPLICATED_PROFILE_PRODUCT("40041"),
     EMPTY_PROFILE_PRODUCT("40042"),
     INVALID_PROFILE_PRODUCT_UPDATE("40043"),
