+++ backend/src/main/java/com/woowacourse/f12/exception/badrequest/InvalidProfileProductCategoryException.java
@@ -1,10 +1,10 @@
 package com.woowacourse.f12.exception.badrequest;
 
-import static com.woowacourse.f12.exception.ErrorCode.PROFILE_CONTAINS_SOFTWARE;
+import static com.woowacourse.f12.exception.ErrorCode.INVALID_CATEGORY_PROFILE_PRODUCT;
 
 public class InvalidProfileProductCategoryException extends InvalidValueException {
 
     public InvalidProfileProductCategoryException() {
-        super(PROFILE_CONTAINS_SOFTWARE, "대표장비로 등록할 수 없는 카테고리입니다.");
+        super(INVALID_CATEGORY_PROFILE_PRODUCT, "대표장비로 등록할 수 없는 카테고리입니다.");
     }
 }
