+++ backend/src/main/java/com/woowacourse/f12/presentation/GlobalExceptionHandler.java
@@ -86,6 +86,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<ExceptionResponse> handleUnhandledException(final Exception e) {
+        e.printStackTrace();
         log.warn(LOG_FORMAT, e.getClass().getSimpleName(), INTERNAL_SERVER_ERROR.getValue(), e.getMessage());
         return ResponseEntity.internalServerError()
                 .body(ExceptionResponse.from(INTERNAL_SERVER_ERROR_MESSAGE, INTERNAL_SERVER_ERROR));
