+++ backend/src/test/java/com/woowacourse/f12/presentation/ControllerTest.java
@@ -1,6 +1,5 @@
 package com.woowacourse.f12.presentation;
 
-import com.woowacourse.f12.RestDocsConfig;
 import com.woowacourse.f12.application.auth.JwtProvider;
 import com.woowacourse.f12.support.AuthTokenExtractor;
 import org.junit.jupiter.api.extension.ExtendWith;
