+++ backend/src/test/java/com/woowacourse/f12/presentation/RestDocsConfig.java
@@ -1,4 +1,4 @@
-package com.woowacourse.f12;
+package com.woowacourse.f12.presentation;
 
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
